INC-802: Moved Prison API classes under `dto.prisonapi` package
This is not strictly necessary but arguably makes much more clear what
classes are tied with Prison API constract and which classes are more under
our control.

Those classes were previously under `service` package despite the fact
they were not services.

Although this move out of the `service` package means imports had to be
adjusted (previously `PrisonApiService` had access to those for free as
they were in the same package). Again, arguably this is a good thing
as the import explicititly tells you now that:
1. these are DTO classes
2. these are classes for responses/requests coming from Prison API
